Improve hooks
To make the hook node more reliable we coerce every functional component to a fragment. By doing this we guarantee that the node on which we add the hooks is consistent between renders. This is important because the dom structure may change but the actual component that generated that dom remained the same.